Rename the Deployments service to Deployment\Backend
Backend says what the abstraction is — a swappable build-dispatch
backend — and Orchestrator names the jobs-service strategy after the
system it talks to, freeing 'Jobs' for the worker that consumes its
callbacks.
