+++ src/Appwrite/Platform/Modules/Functions/Http/Deployments/Duplicate/Create.php
@@ -2,12 +2,12 @@
 
 namespace Appwrite\Platform\Modules\Functions\Http\Deployments\Duplicate;
 
+use Appwrite\Deployment\Backend;
 use Appwrite\Event\Event;
 use Appwrite\Extend\Exception;
 use Appwrite\SDK\AuthType;
 use Appwrite\SDK\Method;
 use Appwrite\SDK\Response as SDKResponse;
-use Appwrite\Service\Deployments;
 use Appwrite\Utopia\Response;
 use Utopia\Database\Database;
 use Utopia\Database\Helpers\ID;
@@ -73,7 +73,7 @@ class Create extends Action
         Response $response,
         Database $dbForProject,
         Event $queueForEvents,
-        Deployments $deployments,
+        Backend $deployments,
         Device $deviceForFunctions,
     ) {
         $function = $dbForProject->getDocument('functions', $functionId);
